refactor: extract anthropicProvider to provider_anthropic.go
